requirements.txt
pandas
kafka-python
configparser
flask
psycopg2
